DPMNeal3

DPMNeal3 provides an interface for Dirichlet process mixture (DPM) models in Julia using Neal's algorithm 3 [1].

Getting Started

Installation

Install with the Julia package manager Pkg, just like any other registered Julia package:

# Press ']' to enter the Pkg REPL mode.
pkg> add DPMNeal3

or

julia> using Pkg; 
julia> Pkg.add("DPMNeal3")

Initializing a Normal DPM

Let us initialize a Normal DPM with N = 1000 observations.

The first step is to set up the environment:

julia> using Random, DPMNeal3
julia> rng = MersenneTwister(1)

Then, we create the generic and specific blocks of this DPM GenericBlock and SpecificBlock, respectively:

julia> gb = GenericBlock(rng, 1000)
julia> sb = SpecificBlock(NormalDPM())

The generic block contains the components that are common to all DPMs (e.g. the cluster labels), while the specific block contains the components that are specific to Normal DPMs (e.g., the NIG kernel hyperparameters).

Performing one Gibbs update

Now, suppose that our sample is contained in the random vector data. Then we can performe one Gibbs update as follows:

julia> update_chainstate!(rng, sb, gb, data)

Consider a DPM o in its so-called stick-breaking representation:

\[\begin{aligned} y_i | \bm{d}, \bm{\theta} &\stackrel{\text{\tiny \it ind}}{\sim} q(\cdot | \theta_{d_i}) \\ d_i | \bm{w} &\stackrel{\text{\tiny \it iid}}{\sim} \text{Categorical}(\bm{w}), & \forall i &\in \mathcal{N} \equiv \{1, \ldots, N\} \\ w_j &= v_j \prod\nolimits_{z < j} (1 - v_z), & \forall j &\in \mathbb{N} \\ v_j &\stackrel{\text{\tiny \it iid}}{\sim} \text{Beta}(1, \alpha) \\ \theta_j &\stackrel{\text{\tiny \it iid}}{\sim} g_0(\cdot) \end{aligned}\]

DPMNeal3.GenericBlock

DPMNeal3.A

DPMNeal3.K

DPMNeal3.N

DPMNeal3.P

DPMNeal3.logpredlik

DPMNeal3.n

DPMNeal3.α

DPMNeal3.GenericBlock — Type

GenericBlock(rng::AbstractRNG, N::Int; K0::Int = 1, a0 = 2.0, b0 = 4.0)

Initialize the generic block of a DPM with N observations, K0 initial clusters and a Gamma prior distribution (with shape a0 and rate b0) for the DP concentration parameter.

DPMNeal3.A — Method

Return the set of active clusters.

DPMNeal3.K — Method

Return the current number of clusters.

DPMNeal3.N — Method

Return the global sample size.

DPMNeal3.P — Method

Return the set of passive clusters.

DPMNeal3.logpredlik — Method

logpredlik(sb, gb::GenericBlock, data, i, k)

Return $\log(y_i | y_{-i}, d_i = k, d_{-i})$.

DPMNeal3.n — Method

Return the current cluster sizes.

DPMNeal3.α — Method

Return the current DP concentration parameter.

push!(LOAD_PATH,"../src/")

rm -rf build; julia make.jl