Consider a DPM o in its so-called stick-breaking representation:

\[\begin{aligned} y_i | \bm{d}, \bm{\theta} &\stackrel{\text{\tiny \it ind}}{\sim} q(\cdot | \theta_{d_i}) \\ d_i | \bm{w} &\stackrel{\text{\tiny \it iid}}{\sim} \text{Categorical}(\bm{w}), & \forall i &\in \mathcal{N} \equiv \{1, \ldots, N\} \\ w_j &= v_j \prod\nolimits_{z < j} (1 - v_z), & \forall j &\in \mathbb{N} \\ v_j &\stackrel{\text{\tiny \it iid}}{\sim} \text{Beta}(1, \alpha) \\ \theta_j &\stackrel{\text{\tiny \it iid}}{\sim} g_0(\cdot) \end{aligned}\]

DPMNeal3.DPM

DPMNeal3.a0

DPMNeal3.active_clusters

DPMNeal3.b0

DPMNeal3.cluster_labels

DPMNeal3.cluster_sizes

DPMNeal3.dp_mass

DPMNeal3.logpredlik

DPMNeal3.max_cluster_label

DPMNeal3.n_clusters

DPMNeal3.parent

DPMNeal3.passive_clusters

DPMNeal3.@forward_dpm_methods

DPMNeal3.DPM — Type

DPM(rng::AbstractRNG, N; K0 = 1, a0 = 2.0, b0 = 4.0)

Initialize a generic DPM with N observations, K0 initial clusters and a Gamma(a0, b0) prior distribution for the mass parameter.

DPMNeal3.a0 — Method

a0(m::AbstractDPM)

Return the shape parameter in the DP mass parameter's prior distribution.

DPMNeal3.active_clusters — Method

active_clusters(m::AbstractDPM)

Return the set of active clusters.

DPMNeal3.b0 — Method

b0(m::AbstractDPM)

Return the scale parameter in the DP mass parameter's prior distribution.

DPMNeal3.cluster_labels — Method

cluster_labels(m::AbstractDPM)

Return the cluster labels.

DPMNeal3.cluster_sizes — Method

cluster_sizes(m::AbstractDPM)

Return the cluster sizes.

DPMNeal3.dp_mass — Method

dp_mass(m::AbstractDPM)

Return the DP mass parameter.

DPMNeal3.logpredlik — Method

logpredlik(sb::AbstractDPM_SB, gb::DPMGB, data, i, k)

Return $\log(y_i | y_{-i}, d_i = k, d_{-i})$.

DPMNeal3.max_cluster_label — Method

max_cluster_label(m::AbstractDPM)

Return the largest (active) cluster label.

DPMNeal3.n_clusters — Method

n_clusters(m::DPM)

Return the number of active clusters.

DPMNeal3.parent — Method

parent(m::AbstractDPM)

Return the parent DPM.

DPMNeal3.passive_clusters — Method

passive_clusters(m::AbstractDPM)

Return a subset of the passive clusters.

DPMNeal3.@forward_dpm_methods — Macro

@forward_all_dpm_methods T.x

Extend all the methods originally defined on type DPM on type T, which call the relevant functions on the field x.

One Example

struct CustomDPM
    gb::DPM
end
@forward_all_dpm_methods CustomDPM.gb
# Now dp_mass(m::Custom) call dp_mass(m.gb)
# Now cluster_labels(m::Custom) call cluster_labels(m.gb)
# etc.

<!– push!(LOAD_PATH,"../src/") rm -rf build; julia make.jl –>