Consider a DPM o in its so-called stick-breaking representation:

\[\begin{aligned} y_i | \bm{d}, \bm{\theta} &\stackrel{\text{\tiny \it ind}}{\sim} q(\cdot | \theta_{d_i}) \\ d_i | \bm{w} &\stackrel{\text{\tiny \it iid}}{\sim} \text{Categorical}(\bm{w}), & \forall i &\in \mathcal{N} \equiv \{1, \ldots, N\} \\ w_j &= v_j \prod\nolimits_{z < j} (1 - v_z), & \forall j &\in \mathbb{N} \\ v_j &\stackrel{\text{\tiny \it iid}}{\sim} \text{Beta}(1, \alpha) \\ \theta_j &\stackrel{\text{\tiny \it iid}}{\sim} g_0(\cdot) \end{aligned}\]

DPMNeal3.DPM

DPMNeal3.active_clusters

DPMNeal3.cluster_labels

DPMNeal3.cluster_sizes

DPMNeal3.dp_mass

DPMNeal3.logpredlik

DPMNeal3.max_cluster_label

DPMNeal3.n_clusters

DPMNeal3.parent

DPMNeal3.passive_clusters

DPMNeal3.update_hyperpars!

DPMNeal3.update_suffstats!

DPMNeal3.update_suffstats!

DPMNeal3.DPM — Type

DPM(rng::AbstractRNG, N; K0 = 1, a0 = 2.0, b0 = 4.0)

Initialize a generic DPM with N observations, K0 initial clusters and a Gamma(a0, b0) prior distribution for the DP mass parameter.

DPMNeal3.active_clusters — Method

active_clusters(m::AbstractDPM)

Return the set of active clusters.

DPMNeal3.cluster_labels — Method

cluster_labels(m::AbstractDPM)

Return the cluster labels.

DPMNeal3.cluster_sizes — Method

cluster_sizes(m::AbstractDPM)

Return the cluster sizes.

DPMNeal3.dp_mass — Method

dp_mass(m::AbstractDPM)

Return the DP mass parameter, α.

DPMNeal3.logpredlik — Method

logpredlik(m::AbstractDPM, data, i::Int, k::Int)

Return $\log(y_i | y_{-i}, x, d_i = k, d_{-i})$, where yi, xi and di are the response, the covariates (if any) and the cluster label associated with the ith observation. Both xi and y_i should be present in the dataset data.

DPMNeal3.max_cluster_label — Method

max_cluster_label(m::AbstractDPM)

Return the largest (active) cluster label.

DPMNeal3.n_clusters — Method

n_clusters(m::AbstractDPM)

Return the number of active clusters.

DPMNeal3.parent — Method

parent(m::AbstractDPM)

Return the parent DPM of m.

DPMNeal3.passive_clusters — Method

passive_clusters(m::AbstractDPM)

Return a subset of the passive clusters.

DPMNeal3.update_hyperpars! — Method

update_hyperpars!(rng::AbstractRNG, m::AbstractDPM, data)

Update the kernel hyperparameters in m, given the dataset data.

DPMNeal3.update_suffstats! — Method

update_suffstats!(m::AbstractDPM, data, i::Int, k0::Int, k1::Int)

Update the sufficient statistics in m after the cluster label of the ith observation changes from k0 to k1, given the dataset data.

DPMNeal3.update_suffstats! — Method

update_suffstats!(m::AbstractDPM, data, i::Int, k0::Int, k1::Int)

Update the sufficient statistics in m from scratch, given the dataset data.

<!– push!(LOAD_PATH,"../src/") rm -rf build; julia make.jl –>